fix: abstracted go to top button
